requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Requests==2.31.0
